third_party/blink/web_tests/platform/generic/fast/loader/document-with-fragment-url-2-expected.txt

ALERT: First load complete.  URL's current hash is #foo
ALERT: Second load complete.  URL's current hash is 
Click here to go to the top of the page

This page should manually be loaded with "#foo" appended to the URL. When you then click the link, the page should navigate to the url without #foo.
